# static/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AutoOnCall",
    "description": "基于 LangChain 的智能oncall运维系统",
    "version": "1.2.1"
  },
  "paths": {
...
    "/api/alerts/alertmanager": {
      "post": {
        "tags": [
          "AIOps告警接入"
        ],
        "summary": "Ingest Alertmanager Webhook",
        "description": "Ingest Alertmanager webhook payloads and create/update Incident state.",
        "operationId": "ingest_alertmanager_webhook_api_alerts_alertmanager_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "auto_diagnose",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Auto Diagnose"
            }
          },
          {
            "name": "X-AutoOnCall-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Autooncall-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertIngestionResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/alerts": {
      "get": {
        "tags": [
          "AIOps告警接入"
        ],
        "summary": "List Alerts",
        "description": "Return recently ingested normalized alerts.",
        "operationId": "list_alerts_api_alerts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "firing",
                    "resolved",
                    "active",
                    "triggered",
                    "inactive",
                    "ok",
                    "closed"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "service_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "Service Name"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "X-AutoOnCall-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Autooncall-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/alerts/{fingerprint}": {
      "get": {
        "tags": [
          "AIOps告警接入"
        ],
        "summary": "Get Alert",
        "description": "Return one normalized alert by fingerprint.",
        "operationId": "get_alert_api_alerts__fingerprint__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "fingerprint",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256,
              "title": "Fingerprint"
            }
          },
          {
            "name": "X-AutoOnCall-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Autooncall-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlertDetailResponse": {
        "properties": {
          "alert": {
            "$ref": "#/components/schemas/AlertEvent"
          }
        },
        "type": "object",
        "required": [
          "alert"
        ],
        "title": "AlertDetailResponse",
        "description": "Detail response for one normalized alert event."
      },
      "AlertEvent": {
        "properties": {
          "source": {
            "type": "string",
            "maxLength": 64,
            "title": "Source",
            "default": "alertmanager"
          },
          "fingerprint": {
            "type": "string",
            "maxLength": 128,
            "title": "Fingerprint"
          },
          "incident_id": {
            "type": "string",
            "maxLength": 128,
            "title": "Incident Id"
          },
          "status": {
            "type": "string",
            "maxLength": 32,
            "title": "Status",
            "default": "firing"
          },
          "alertname": {
            "type": "string",
            "maxLength": 256,
            "title": "Alertname",
            "default": "UnknownAlert"
          },
          "service_name": {
            "type": "string",
            "maxLength": 128,
            "title": "Service Name",
            "default": "unknown-service"
          },
          "severity": {
            "type": "string",
            "maxLength": 16,
            "title": "Severity",
            "default": "P2"
          },
          "environment": {
            "type": "string",
            "maxLength": 64,
            "title": "Environment",
            "default": "unknown"
          },
          "summary": {
            "type": "string",
            "maxLength": 4096,
            "title": "Summary",
            "default": ""
          },
          "description": {
            "type": "string",
            "maxLength": 4096,
            "title": "Description",
            "default": ""
          },
          "labels": {
            "additionalProperties": true,
            "type": "object",
            "title": "Labels"
          },
          "annotations": {
            "additionalProperties": true,
            "type": "object",
            "title": "Annotations"
          },
          "starts_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts At"
          },
          "ends_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends At"
          },
          "generator_url": {
            "type": "string",
            "maxLength": 2048,
            "title": "Generator Url",
            "default": ""
          },
          "raw_payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Raw Payload"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "fingerprint",
          "incident_id"
        ],
        "title": "AlertEvent",
        "description": "Normalized alert event persisted before it becomes an Incident."
      },
      "AlertIngestionItem": {
        "properties": {
          "event": {
            "$ref": "#/components/schemas/AlertEvent"
          },
          "created": {
            "type": "boolean",
            "title": "Created",
            "default": false
          },
          "deduplicated": {
            "type": "boolean",
            "title": "Deduplicated",
            "default": false
          },
          "previous_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Status"
          },
          "status_changed": {
            "type": "boolean",
            "title": "Status Changed",
            "default": false
          },
          "reopened": {
            "type": "boolean",
            "title": "Reopened",
            "default": false
          },
          "stale_ignored": {
            "type": "boolean",
            "title": "Stale Ignored",
            "default": false
          },
          "incident_id": {
            "type": "string",
            "title": "Incident Id"
          },
          "incident_status": {
            "type": "string",
            "title": "Incident Status"
          },
          "status_reason": {
            "type": "string",
            "title": "Status Reason",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "event",
          "incident_id",
          "incident_status"
        ],
        "title": "AlertIngestionItem",
        "description": "One normalized alert and the Incident lifecycle result it produced."
      },
      "AlertIngestionResult": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source",
            "default": "alertmanager"
          },
          "received": {
            "type": "integer",
            "title": "Received",
            "default": 0
          },
          "created": {
            "type": "integer",
            "title": "Created",
            "default": 0
          },
          "deduplicated": {
            "type": "integer",
            "title": "Deduplicated",
            "default": 0
          },
          "resolved": {
            "type": "integer",
            "title": "Resolved",
            "default": 0
          },
          "stale_ignored": {
            "type": "integer",
            "title": "Stale Ignored",
            "default": 0
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AlertIngestionItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "title": "AlertIngestionResult",
        "description": "Response returned after processing an alert webhook."
      },
      "AlertListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AlertEvent"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "title": "AlertListResponse",
        "description": "List response for normalized alert events."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}